Port extension to use WebExtensions API when available
While the extension did happen to work as-is on Firefox thanks to their
compatibility API, it wasn't much work to just make it use the
WebExtensions API as nearly nothing changes.

This also refactors the code to be strict mode, avoid leaking variables
to global scope, and to use the native clipboard API, if available,
instead of manipulating the page's text selection. The item id mapping
loop was optimized as well.